@@ -21,10 +21,16 @@
         "build": "encore production --progress"
     },
     "dependencies": {
-        "axios": "^0.18.0",
-        "react": "^16.8.6",
-        "react-custom-roulette": "^1.2.0",
-        "react-dom": "^16.8.6",
-        "react-router-dom": "^5.0.0"
+        "@emotion/react": "^11.11.1",
+        "@emotion/styled": "^11.11.0",
+        "@fontsource/roboto": "^5.0.8",
+        "@mui/icons-material": "^5.14.8",
+        "@mui/material": "^5.14.8",
+        "@mui/x-date-pickers": "^6.13.0",
+        "dayjs": "^1.11.9",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
+        "react-query": "^3.39.3",
+        "react-router-dom": "^6.15.0"
     }
 }
