@@ -1,11 +1,12 @@
-import React from 'react';
+import React from "react";
 
 const AccueilPage = (props) => {
-    return ( 
-    <h1>
-        Titouan COLLOT test
-    </h1>
-    );
-}
- 
+  return (
+      <h3>
+
+      </h3>
+    
+  );
+};
+
 export default AccueilPage;
